Move notebook directory utils file
